+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Next Release
 
+## API changes
+
+PR [#559](https://github.com/IAMconsortium/pyam/pull/559) marked
+the attribute `_LONG_IDX` as deprecated. Please use `dimensions` instead. 
+
+## Individual updates
+
+- [#559](https://github.com/IAMconsortium/pyam/pull/559) Add attribute dimensions, fix compatibility with pandas v1.3
 - [#557](https://github.com/IAMconsortium/pyam/pull/557) Swap time for year keeping subannual resolution
 - [#556](https://github.com/IAMconsortium/pyam/pull/556) Set explicit minimum numpy version (1.19.0)